Rename module to not clash in py2
